backend/run.py
from os import path

from todo_list_api import create_app
from dotenv import load_dotenv

basedir = path.abspath(path.dirname(__file__))
load_dotenv(path.join(basedir, '.dev.env'))

app = create_app()


