backend/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
contextlib2==21.6.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-PyMongo==2.3.0
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
packaging==21.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.4
pytest-schema==0.1.0
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
schema==0.7.4
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
